abstract away content component, add interface, remove old code comments
