+++ src/containers/homepage-container.tsx
@@ -1,20 +1,7 @@
-import {
-  Flex,
-  Input,
-  Select,
-  Typography,
-  Row,
-  Col,
-  Image,
-  Anchor,
-  Space,
-  message,
-} from 'antd';
-import Card from 'antd/es/card/Card';
+import { message } from 'antd';
 
 import { useState } from 'react';
 import { debounce } from '../utils/debounce.utils';
-import Meta from 'antd/es/card/Meta';
 import { IRepository, IUserProfile } from '../utils/interfaces.utils';
 import {
   fetchUserFollowers,
@@ -23,8 +10,8 @@ import {
 } from '../services/github.service';
 import getRandomColor from '../utils/randomColor.util';
 import { SearchInputComponent } from '../app/components/common/searchInput/searchInput';
+import { ContentComponent } from '../app/components/content/content.component';
 
-const { Title } = Typography;
 export default function Search() {
   const [username, setUsername] = useState('');
   const [userProfile, setUserProfile] = useState<IUserProfile[]>([]);
@@ -89,104 +76,25 @@ export default function Search() {
     setSearchedUsers([...searchedUsers, username]);
   };
 
-  const toggleReposCard = async (username: string) => {
-    setIsRepoExpanded((prevState) => !prevState);
-
-    if (!isRepoExpanded) {
-      const repos = await fetchUserRepos(username);
-      setExpandedUserRepos(repos);
-    }
-  };
-
   const debouncedProfileSearch = debounce(searchUser, 1000);
 
   return (
-    <Flex vertical gap='middle' wrap='wrap'>
+    <>
       <SearchInputComponent
         username={username}
         setUsername={setUsername}
         debouncedProfile={debouncedProfileSearch}
       />
 
-      {userProfile.length > 0 && (
-        <Row gutter={[182, 8]}>
-          {userProfile?.map((profile, index) => (
-            <Col key={index} span={8}>
-              {profile.login !== undefined && (
-                <Card
-                  onClick={() => {
-                    toggleReposCard(profile.login);
-                    setActiveColor(profile.background as string);
-                  }}
-                  hoverable
-                  style={{
-                    width: 240,
-                    backgroundColor: profile.background,
-                  }}
-                  cover={<Image alt='user dp' src={profile.avatar_url} />}
-                >
-                  <Meta
-                    title={profile.login}
-                    description={
-                      <Anchor
-                        items={[
-                          {
-                            key: 'profile_url',
-                            href: profile.html_url,
-                            title: profile.login,
-                          },
-                        ]}
-                      />
-                    }
-                  />
-
-                  <div>
-                    <Title level={5}>
-                      Followers: {profile?.followers?.length ?? 0}
-                    </Title>
-                  </div>
-                </Card>
-              )}
-            </Col>
-          ))}
-        </Row>
-      )}
-
-      {!isRepoExpanded ? null : (
-        <Row gutter={[182, 16]}>
-          {expandedUserRepos?.map((repo, index) => (
-            <Col key={index} span={8}>
-              {repo.name !== undefined && (
-                <Card
-                  hoverable
-                  style={{
-                    width: 240,
-                    background: activeColor,
-                  }}
-                >
-                  <Meta
-                    title={repo.name}
-                    description={
-                      <Anchor
-                        items={[
-                          {
-                            key: 'profile_url',
-                            href: repo.html_url,
-                            title: repo.name,
-                          },
-                        ]}
-                      />
-                    }
-                  />
-                  <div>
-                    <Title level={5}>Stars: {repo.stargazers_count}</Title>
-                  </div>
-                </Card>
-              )}
-            </Col>
-          ))}
-        </Row>
-      )}
-    </Flex>
+      <ContentComponent
+        userProfile={userProfile}
+        isRepoExpanded={isRepoExpanded}
+        setIsRepoExpanded={setIsRepoExpanded}
+        expandedUserRepos={expandedUserRepos}
+        setExpandedUserRepos={setExpandedUserRepos}
+        activeColor={activeColor}
+        setActiveColor={setActiveColor}
+      />
+    </>
   );
 }
